+++ printf/ft_printf_utils2.c
@@ -6,14 +6,14 @@
 /*   By: ilyanar <ilyanar@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 22:12:15 by ilyanar           #+#    #+#             */
-/*   Updated: 2023/11/13 22:40:45 by ilyanar          ###   ########.fr       */
+/*   Updated: 2023/11/14 05:16:09 by ilyanar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft/libft.h"
 
-void	putnbr_u(long long nb, int *n)
+void	putnbr_u(unsigned int nb, int *n)
 {
 	if (nb < 0)
 		nb *= -1;
